Docs : Trip 파일 구조 변경 2
 - 생산성 향상을 위한 파일구조 도메인 단위로 리팩토링

  Fixed #99
