+++ src/main/kotlin/kr/kro/tripsketch/comment/CommentRepository.kt
@@ -1,6 +1,6 @@
-package kr.kro.tripsketch.repositories
+package kr.kro.tripsketch.comment
 
-import kr.kro.tripsketch.domain.Comment
+import kr.kro.tripsketch.comment.Comment
 import org.springframework.data.mongodb.repository.MongoRepository
 import org.springframework.stereotype.Repository
