+++ src/main/kotlin/kr/kro/tripsketch/comment/dtos/CommentDto.kt
@@ -1,4 +1,4 @@
-package kr.kro.tripsketch.dto
+package kr.kro.tripsketch.comment.dtos
 
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.Size
