+++ src/main/kotlin/kr/kro/tripsketch/trip/TripRepository.kt
@@ -1,6 +1,6 @@
-package kr.kro.tripsketch.repositories
+package kr.kro.tripsketch.trip
 
-import kr.kro.tripsketch.domain.Trip
+import kr.kro.tripsketch.trip.Trip
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.mongodb.repository.MongoRepository
