+++ src/main/kotlin/kr/kro/tripsketch/trip/TripService.kt
@@ -1,13 +1,16 @@
-package kr.kro.tripsketch.services
+package kr.kro.tripsketch.trip
 
-import kr.kro.tripsketch.domain.HashtagInfo
-import kr.kro.tripsketch.domain.Trip
 import kr.kro.tripsketch.dto.*
 import kr.kro.tripsketch.exceptions.ForbiddenException
-import kr.kro.tripsketch.repositories.CommentRepository
+import kr.kro.tripsketch.comment.CommentRepository
+import kr.kro.tripsketch.comment.CommentService
+import kr.kro.tripsketch.comment.dtos.CommentDto
 import kr.kro.tripsketch.repositories.FollowRepository
-import kr.kro.tripsketch.repositories.TripRepository
 import kr.kro.tripsketch.repositories.UserRepository
+import kr.kro.tripsketch.services.ImageService
+import kr.kro.tripsketch.services.NotificationService
+import kr.kro.tripsketch.services.UserService
+import kr.kro.tripsketch.trip.dtos.*
 import kr.kro.tripsketch.utils.EnvLoader
 import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
