+++ src/main/kotlin/kr/kro/tripsketch/trip/dtos/TripAndCommentResponseDto.kt
@@ -1,4 +1,6 @@
-package kr.kro.tripsketch.dto
+package kr.kro.tripsketch.trip.dtos
+
+import kr.kro.tripsketch.comment.dtos.CommentDto
 
 /**
  * 여행 정보와 댓글들의 목록을 담은 응답 DTO 클래스입니다.
