+++ src/main/kotlin/kr/kro/tripsketch/trip/dtos/TripCardDto.kt
@@ -1,4 +1,4 @@
-package kr.kro.tripsketch.dto
+package kr.kro.tripsketch.trip.dtos
 
 import java.time.LocalDateTime
