+++ src/main/kotlin/kr/kro/tripsketch/trip/dtos/TripLikesDto.kt
@@ -1,4 +1,4 @@
-package kr.kro.tripsketch.dto
+package kr.kro.tripsketch.trip.dtos
 
 data class TripLikesDto(
     var id: String? = null,
